refact(plugins) split client/host requires on all plugins
* smaller client/host bundles
* consistent with iframe-resizer
